splited index.ts
